requirements.txt
coqui-tts==0.26.2
torch>=2.2.0
torchaudio>=2.2.0
numpy>=1.24.0
pygame==2.5.2
customtkinter==5.2.1
coqpit>=0.0.16
fsspec>=2023.1.0
huggingface-hub>=0.19.4
phonemizer>=3.0.1
unidecode>=1.3.6
pypinyin>=0.49.0
jieba>=0.42.1
inflect>=7.0.0
eng_to_ipa>=0.0.2
num2words>=0.5.12
espeak-ng==0.1.3
 